Fix Windows parallel-build race and skip flaky CI tests
Use OBJECT libraries for extension proto-generated code so that
static and shared plugin targets share a single compilation, avoiding
the MSBuild race where both targets try to generate the same .pb.cc
file simultaneously ("Permission denied"). Applied to acorn-rtc,
acorn-scsi, and test-scratch-ram extensions.

Widen PacingController::target_clock_hz_ from uint32_t to uint64_t
to match PacingConfig::base_clock_hz and eliminate MSVC C4244
truncation warning.

Skip two timing-sensitive TypeScript integration tests in CI (detected
via the CI environment variable) that are unreliable on slow GitHub
Actions runners but pass locally on faster hardware.

+++ src/core/include/beebium/PacingController.hpp
@@ -39,7 +39,7 @@ public:
     /// @param max_ratio        Maximum cycles as multiple of nominal per-quantum
     ///                         cycles. 1.015 = allow 1.5% overshoot for smooth
     ///                         catch-up after deficits.
-    PacingController(uint32_t target_clock_hz, int64_t quantum_ns,
+    PacingController(uint64_t target_clock_hz, int64_t quantum_ns,
                      double max_ratio = 3.0)
         : target_clock_hz_(target_clock_hz)
         , nominal_cycles_(static_cast<uint32_t>(
@@ -76,7 +76,7 @@ public:
     uint32_t max_cycles() const { return max_cycles_; }
 
 private:
-    uint32_t target_clock_hz_;
+    uint64_t target_clock_hz_;
     uint32_t nominal_cycles_;   // cycles per quantum at 1x speed
     uint32_t max_cycles_;       // ceiling per tick
